Moved acl related code to a new class ACL.
